modificação do método CalculaValorTotal, modificação do método CalculaTotalDeImposto, criação do método RemoverItems, criação do método EditaQuantidadeDeItems, criação do método Adicionar, criação do método Modificar, modificações em alguns tipos de propriedades, mais correções de testes que foram quebrados.

+++ Atividade-Wiz-Semana8/Comex.Models/ItemsDoPedido.cs
@@ -11,13 +11,14 @@ namespace Comex.Models
     {
         public readonly Produto Produto;
         public int Quantidade { get; set; }
+
         public readonly decimal Total;
 
         public ItemsDoPedido(Produto produto, int quantidade)
         {
             Produto = produto;
             Quantidade = quantidade;
-            Total = (decimal)(Produto.PrecoUnitario * Quantidade);
+            Total = Produto.PrecoUnitario * Quantidade;
         }
     }
 }
